ensure authorization and test endpoints

+++ app/Http/Controllers/Api/V1/Project/ProjectController.php
@@ -43,8 +43,7 @@ class ProjectController extends Controller
     public function store(CreateProjectRequest $request): JsonResponse
     {
         $project = $this->service->create(
-//            owner: $request->user(),
-            owner: $this->resolveUser($request),
+            owner: $request->user(),
             data:  $request->validated(),
         );
 
@@ -61,7 +60,7 @@ class ProjectController extends Controller
     public function show(Request $request, string $id): JsonResponse
     {
         $project = $this->service->show($id);
-        //        $this->authorize('view', $project);
+        $this->authorize('view', $project);
 
         return response()->json([
             'data' => new ProjectResource($project),
@@ -74,7 +73,7 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project): JsonResponse
     {
-//        $this->authorize('update', $project);
+        $this->authorize('update', $project);
 
         $updated = $this->service->update($project, $request->validated());
 
@@ -90,7 +89,7 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project): JsonResponse
     {
-//        $this->authorize('delete', $project);
+        $this->authorize('delete', $project);
 
         $this->service->delete($project);
